+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/update/CayenneUpdateProcessorFactoryFactoryProvider.java
@@ -42,7 +42,8 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
             @Inject CayenneCreatedResponseStage createdResponseStage,
             @Inject CayenneCreatedOrOkResponseStage createdOrOkResponseStage,
 
-            @Inject FilterResultStage filterResultStage
+            @Inject FilterResultStage filterResultStage,
+            @Inject EncoderInstallStage encoderInstallStage
     ) {
 
         this.createStages = new EnumMap<>(UpdateStage.class);
@@ -56,6 +57,7 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
         this.createStages.put(UpdateStage.COMMIT, commitStage);
         this.createStages.put(UpdateStage.FILL_RESPONSE, createdResponseStage);
         this.createStages.put(UpdateStage.FILTER_RESULT, filterResultStage);
+        this.createStages.put(UpdateStage.ENCODE, encoderInstallStage);
 
         this.updateStages = new EnumMap<>(UpdateStage.class);
         this.updateStages.put(UpdateStage.START, startStage);
@@ -68,6 +70,7 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
         this.updateStages.put(UpdateStage.COMMIT, commitStage);
         this.updateStages.put(UpdateStage.FILL_RESPONSE, okResponseStage);
         this.updateStages.put(UpdateStage.FILTER_RESULT, filterResultStage);
+        this.updateStages.put(UpdateStage.ENCODE, encoderInstallStage);
 
         this.createOrUpdateStages = new EnumMap<>(UpdateStage.class);
         this.createOrUpdateStages.put(UpdateStage.START, startStage);
@@ -80,6 +83,7 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
         this.createOrUpdateStages.put(UpdateStage.COMMIT, commitStage);
         this.createOrUpdateStages.put(UpdateStage.FILL_RESPONSE, createdOrOkResponseStage);
         this.createOrUpdateStages.put(UpdateStage.FILTER_RESULT, filterResultStage);
+        this.createOrUpdateStages.put(UpdateStage.ENCODE, encoderInstallStage);
 
         this.idempotentCreateOrUpdateStages = new EnumMap<>(UpdateStage.class);
         this.idempotentCreateOrUpdateStages.put(UpdateStage.START, startStage);
@@ -92,6 +96,7 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
         this.idempotentCreateOrUpdateStages.put(UpdateStage.COMMIT, commitStage);
         this.idempotentCreateOrUpdateStages.put(UpdateStage.FILL_RESPONSE, createdOrOkResponseStage);
         this.idempotentCreateOrUpdateStages.put(UpdateStage.FILTER_RESULT, filterResultStage);
+        this.idempotentCreateOrUpdateStages.put(UpdateStage.ENCODE, encoderInstallStage);
 
         this.idempotentFullSyncStages = new EnumMap<>(UpdateStage.class);
         this.idempotentFullSyncStages.put(UpdateStage.START, startStage);
@@ -104,6 +109,7 @@ public class CayenneUpdateProcessorFactoryFactoryProvider implements Provider<Up
         this.idempotentFullSyncStages.put(UpdateStage.COMMIT, commitStage);
         this.idempotentFullSyncStages.put(UpdateStage.FILL_RESPONSE, createdOrOkResponseStage);
         this.idempotentFullSyncStages.put(UpdateStage.FILTER_RESULT, filterResultStage);
+        this.idempotentFullSyncStages.put(UpdateStage.ENCODE, encoderInstallStage);
     }
 
     @Override