+++ agrest-engine/src/main/java/io/agrest/UpdateStage.java
@@ -44,5 +44,12 @@ public enum UpdateStage {
      *
      * @since 4.8
      */
-    FILTER_RESULT
+    FILTER_RESULT,
+
+    /**
+     * A stage when the result is encoded, or encoders are installed, depending on pipeline implementation
+     *
+     * @since 4.8
+     */
+    ENCODE
 }