+++ agrest-engine/src/main/java/io/agrest/runtime/processor/select/ApplyServerParamsStage.java
@@ -5,7 +5,6 @@ import io.agrest.encoder.EntityEncoderFilter;
 import io.agrest.processor.Processor;
 import io.agrest.processor.ProcessorOutcome;
 import io.agrest.runtime.constraints.IConstraintsHandler;
-import io.agrest.runtime.encoder.IEncoderService;
 import org.apache.cayenne.di.Inject;
 
 import java.util.ArrayList;
@@ -17,16 +16,13 @@ import java.util.List;
 public class ApplyServerParamsStage implements Processor<SelectContext<?>> {
 
     private IConstraintsHandler constraintsHandler;
-    private IEncoderService encoderService;
     private List<EntityEncoderFilter> filters;
 
     public ApplyServerParamsStage(
             @Inject IConstraintsHandler constraintsHandler,
-            @Inject IEncoderService encoderService,
             @Inject List<EntityEncoderFilter> filters) {
 
         this.constraintsHandler = constraintsHandler;
-        this.encoderService = encoderService;
         this.filters = filters;
     }
 
@@ -37,21 +33,11 @@ public class ApplyServerParamsStage implements Processor<SelectContext<?>> {
     }
 
     protected <T> void doExecute(SelectContext<T> context) {
-
-        ResourceEntity<T> entity = context.getEntity();
-
-        constraintsHandler.constrainResponse(entity, context.getSizeConstraints(), context.getConstraint());
-
+        constraintsHandler.constrainResponse(context.getEntity(), context.getSizeConstraints(), context.getConstraint());
         initEncoderFilters(context);
-
-        // make sure we create the encoder, even if we end up with an empty
-        // list, as we need to encode the totals
-
-        if (context.getEncoder() == null) {
-            context.setEncoder(encoderService.dataEncoder(entity));
-        }
     }
 
+    @Deprecated
     protected void initEncoderFilters(SelectContext<?> context) {
         List<EntityEncoderFilter> filters = mergeFilters(context.getEntityEncoderFilters());
         if (!filters.isEmpty()) {
@@ -59,6 +45,7 @@ public class ApplyServerParamsStage implements Processor<SelectContext<?>> {
         }
     }
 
+    @Deprecated
     protected void initEncoderFilters(ResourceEntity<?> entity, List<EntityEncoderFilter> filters) {
 
         for (EntityEncoderFilter filter : filters) {
@@ -72,6 +59,7 @@ public class ApplyServerParamsStage implements Processor<SelectContext<?>> {
         }
     }
 
+    @Deprecated
     protected List<EntityEncoderFilter> mergeFilters(List<EntityEncoderFilter> requestFilters) {
 
         if (requestFilters == null || requestFilters.isEmpty()) {