+++ agrest-engine/src/main/java/io/agrest/runtime/processor/select/SelectProcessorFactoryProvider.java
@@ -22,7 +22,8 @@ public class SelectProcessorFactoryProvider implements Provider<SelectProcessorF
             @Inject ApplyServerParamsStage applyServerParamsStage,
             @Inject AssembleQueryStage assembleQueryStage,
             @Inject FetchDataStage fetchDataStage,
-            @Inject FilterResultStage filterResultStage) {
+            @Inject FilterResultStage filterResultStage,
+            @Inject EncoderInstallStage encoderInstallStage) {
 
         stages = new EnumMap<>(SelectStage.class);
         stages.put(SelectStage.START, startStage);
@@ -32,6 +33,7 @@ public class SelectProcessorFactoryProvider implements Provider<SelectProcessorF
         stages.put(SelectStage.ASSEMBLE_QUERY, assembleQueryStage);
         stages.put(SelectStage.FETCH_DATA, fetchDataStage);
         stages.put(SelectStage.FILTER_RESULT, filterResultStage);
+        stages.put(SelectStage.ENCODE, encoderInstallStage);
     }
 
     @Override